fix manage project script
