+++ package.json
@@ -10,6 +10,7 @@
     "prepare-open-api-for-all-published-sdks": "npm run prepare-open-api-for-sdk -- --language=ruby && npm run prepare-open-api-for-sdk -- --language=java && npm run prepare-open-api-for-sdk -- --language=php",
     "manage-project": "ts-node scripts/build-update-md-tables-from-openapi.ts && ts-node ./scripts/manage-project.ts",
     "fix-schemas-with-refs": "ts-node ./scripts/fix-schemas-with-refs.ts",
+    "create-open-api-version-to-upload": "ts-node ./scripts/create-open-api-version-to-upload.ts",
     "fix-schemas-properties-with-single-enum": "ts-node ./scripts/fix-schemas-properties-with-single-enum.ts",
     "readme-upload-missing-images": "ts-node ./scripts/readme-upload-missing-images.ts",
     "count-important-statistics-about-openapi": "ts-node ./scripts/count-important-statistics-about-openapi.ts",
@@ -26,7 +27,7 @@
     "test-java-sdk": "concurrently \"cd ./sdks/java && docker build --no-cache -t java .\"",
     "test-ruby-sdk": "concurrently \"cd ./sdks/ruby && docker build --no-cache -t ruby . && docker run --rm ruby\"",
     "test": "jest",
-    "pre-commit": "npm run generate-endpoints-coverage-doc -- --generateFor=default;npm run fix-schemas-properties-with-single-enum && npm run fix-schemas-with-refs && npm run build-update-md-tables-from-openapi && npm run prepare-open-api-for-all-published-sdks && npm run build-production-openapi",
+    "pre-commit": "npm run generate-endpoints-coverage-doc -- --generateFor=default;npm run fix-schemas-properties-with-single-enum && npm run fix-schemas-with-refs && npm run create-open-api-version-to-upload && npm run build-update-md-tables-from-openapi && npm run prepare-open-api-for-all-published-sdks && npm run build-production-openapi",
     "validate-openapi": "npx @openapitools/openapi-generator-cli validate -i ./production/readOnly-openAPI.json && npx @openapitools/openapi-generator-cli validate -i ./reference/OpenAPIWebhooks.json"
   },
   "dependencies": {
