+++ scripts/manage-project.ts
@@ -45,8 +45,13 @@ const main = async ({
   create?: boolean;
   update?: boolean;
 }) => {
-  await createOpenAPIVersionToUpload();
-  return;
+  const { error } = await runCliProcess({
+    command: `npm run create-open-api-version-to-upload`,
+    resolveErrorAsFalse: true,
+  });
+  if (error) {
+    console.log(colors.red("Could not build openApi file to upload!"));
+  }
   if (!(await validate({ help, version, create, update }))) {
     return;
   }
@@ -73,7 +78,6 @@ const uploadOpenApiFileWithMaxNumberOfAttempts = async (
   version,
   maxNumberOfUploadingAttempts = 3,
 ) => {
-  await createOpenAPIVersionToUpload();
   console.log(
     colors.green(
       "UPLOADING OPEN API FILE... PLEASE WAIT... THIS MAY TAKE UP TO A MINUTE",
@@ -129,50 +133,6 @@ const isVersionExists = async (version: string) => {
         },
       )
     ).status === 200
-  );
-};
-
-const createOpenAPIVersionToUpload = async () => {
-  console.log(
-    colors.green("CREATING OPEN API FILE TO UPLOAD... PLEASE WAIT..."),
-  );
-
-  const openApiPath = path.join(__dirname, "../reference/OpenAPI.json");
-  const openAPIContent = JSON.parse(
-    (await fsPromises.readFile(openApiPath)).toString(),
-  );
-
-  const pathToTmp = path.join(__dirname, "../tmp");
-  if (!fs.existsSync(pathToTmp)) {
-    fs.mkdirSync(pathToTmp);
-  }
-
-  const pathToTmpReferenceToUpload = path.join(
-    __dirname,
-    "../tmp/referenceToUpload",
-  );
-  if (!fs.existsSync(pathToTmpReferenceToUpload)) {
-    fs.mkdirSync(pathToTmpReferenceToUpload);
-  }
-
-  let newOpenApiFile = {
-    ...openAPIContent,
-    components: {
-      ...openAPIContent.components,
-      schemas: removeAdditionalPropertiesFromSchemas(
-        openAPIContent.components.schemas,
-        ["ExportsCreateRequestBody", "ExportsCreateResponseBody"],
-      ),
-    },
-  };
-  newOpenApiFile.openapi = "3.1.0";
-
-  //replaceTitles
-  newOpenApiFile = readmeReplaceTitle(newOpenApiFile);
-
-  await fsPromises.writeFile(
-    path.join(__dirname, "../tmp/referenceToUpload/OpenAPI.json"),
-    JSON.stringify(newOpenApiFile, null, 2),
   );
 };
